i added include entity class.And i changed property type.

+++ EcommerceService.Core/DTOs/Campaign/CreateCampaignInput.cs
@@ -5,6 +5,6 @@ public class CreateCampaignInput
     public string Name { get; set; }
     public string ProductCode { get; set; }
     public int Duration { get; set; }
-    public double PriceManipulationLimit { get; set; }
+    public decimal PriceManipulationLimit { get; set; }
     public int TargetSalesCount { get; set; }
 }
